% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_variations_rds.R
\name{check_list_has_valid_snp_ids}
\alias{check_list_has_valid_snp_ids}
\title{Check if all the SNP IDs in the list are in the tibble of SNP IDs}
\usage{
check_list_has_valid_snp_ids(tibbles, t_snp_ids)
}
\arguments{
\item{tibbles}{List of tibbles with snp_id and variation}

\item{t_snp_ids}{tibble with only the column \code{snp_id}}
}
\description{
Check if all the SNP IDs in the list are in the tibble of SNP IDs
}
