% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topo_files.R
\name{check_topo_files}
\alias{check_topo_files}
\title{Check the \code{.topo} files for being valid.
Will \link{stop} if not}
\usage{
check_topo_files(topo_filenames)
}
\arguments{
\item{topo_filenames}{name of one or more \code{.topo} files
with protein topologies}
}
\description{
Check the \code{.topo} files for being valid.
Will \link{stop} if not
}
