% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fasta_file.R
\name{create_fasta_file}
\alias{create_fasta_file}
\title{Create a FASTA file with the sequences for each variation}
\usage{
create_fasta_file(variations_csv_filename, verbose = FALSE)
}
\arguments{
\item{variations_csv_filename}{name of a \code{[gene_name]_variations.csv}
file}

\item{verbose}{set to TRUE for more output}
}
\description{
Create a FASTA file with the sequences for each variation
}
