% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fasta_files.R
\name{create_fasta_files}
\alias{create_fasta_files}
\title{Create a FASTA file with the sequences for each variation}
\usage{
create_fasta_files(variations_csv_filenames, verbose = FALSE)
}
\arguments{
\item{variations_csv_filenames}{names of one or more
\code{[gene_name]_variations.csv}
files}

\item{verbose}{set to TRUE for more output}
}
\description{
Create a FASTA file with the sequences for each variation
}
