% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_name_snps.R
\name{create_gene_name_snps}
\alias{create_gene_name_snps}
\title{Per gene name, save a table with SNP IDs to file}
\usage{
create_gene_name_snps(gene_names_filename, folder_name)
}
\arguments{
\item{gene_names_filename}{the filename to save
the gene IDs and gene names to.
For the experiment, use \code{gene_names.csv}}
}
\value{
all names of the files created
}
\description{
Per gene name, save a table with SNP IDs to file
}
