% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_name_snps.R,
%   R/create_gene_name_snps_files.R
\name{create_gene_name_snps_files}
\alias{create_gene_name_snps_files}
\title{Per gene name, save a table with SNP IDs to file}
\usage{
create_gene_name_snps_files(gene_names_filename)

create_gene_name_snps_files(gene_names_filename)
}
\arguments{
\item{gene_names_filename}{the filename to save
the gene IDs and gene names to.
For the experiment, use \code{gene_names.csv}}
}
\value{
all names of the files created

all names of the files created
}
\description{
Per gene name, save a table with SNP IDs to file

Per gene name, save a table with SNP IDs to file
}
