% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_names.R
\name{create_gene_names}
\alias{create_gene_names}
\title{Create a table with gene ID and gene name}
\usage{
create_gene_names(gene_ids_filename, gene_names_filename)
}
\arguments{
\item{gene_ids_filename}{the filename to save
the gene IDs to.
For the experiment, use \code{gene_ids.csv}}

\item{gene_names_filename}{the filename to save
the gene IDs and gene names to.
For the experiment, use \code{gene_names.csv}}
}
\description{
Create a table with gene ID and gene name
}
