% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_is_in_tmh_file.R
\name{create_is_in_tmh_file}
\alias{create_is_in_tmh_file}
\title{Create one \code{[gene_name]_is_in_tmh.csv} file.}
\usage{
create_is_in_tmh_file(variations_csv_filename, topo_filename)
}
\arguments{
\item{variations_csv_filename}{name of a \code{[gene_name]_variations.csv}
file}

\item{topo_filename}{name of a \code{.topo} file with a protein topology}
}
\description{
Create one \code{[gene_name]_is_in_tmh.csv} file.
}
