% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_is_in_tmh_files.R
\name{create_is_in_tmh_files}
\alias{create_is_in_tmh_files}
\title{Create the \code{[gene_name]_is_in_tmh.csv} files.}
\usage{
create_is_in_tmh_files(topo_filenames, verbose = FALSE)
}
\arguments{
\item{topo_filenames}{name of one or more \code{.topo} files
with protein topologies}

\item{verbose}{set to TRUE for more output}
}
\description{
Create the \code{[gene_name]_is_in_tmh.csv} files.
}
