% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_results_file.R
\name{create_results_file}
\alias{create_results_file}
\title{Create the results file}
\usage{
create_results_file(is_in_tmh_filenames)
}
\arguments{
\item{is_in_tmh_filenames}{the filename to save
the \code{is_in_tmh} tables to (each of which can be read by
\link{read_is_in_tmh_file}).
For the experiment, use \code{[gene_name]_is_in_tmh.csv}}
}
\description{
Create the results file
}
