% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_variations_csv_file.R
\name{create_snp_variations_csv_file}
\alias{create_snp_variations_csv_file}
\title{For each \code{rds} file
(each containing a list of tibbles with SNP ID and variations),
convert it to a \code{csv} file with the merged table}
\usage{
create_snp_variations_csv_file(variations_rds_filename)
}
\value{
names of the \code{csv} files created
}
\description{
For each \code{rds} file
(each containing a list of tibbles with SNP ID and variations),
convert it to a \code{csv} file with the merged table
}
