% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_variations_csv_files.R
\name{create_snp_variations_csv_files}
\alias{create_snp_variations_csv_files}
\title{For each \code{rds} file
(each containing a list of tibbles with SNP ID and variations),
convert it to a \code{csv} file with the merged table}
\usage{
create_snp_variations_csv_files(variations_rds_filenames, verbose = FALSE)
}
\arguments{
\item{variations_rds_filenames}{names of one or more
\code{[gene_name]_variations.rds}
files}

\item{verbose}{set to TRUE for more output}
}
\value{
names of the \code{csv} files created
}
\description{
For each \code{rds} file
(each containing a list of tibbles with SNP ID and variations),
convert it to a \code{csv} file with the merged table
}
