% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_variations_rds.R
\name{create_snp_variations_rds}
\alias{create_snp_variations_rds}
\title{For each gene,
create an \code{rds} file containing a tibble of variations per SNP}
\usage{
create_snp_variations_rds(gene_names_filename)
}
\arguments{
\item{gene_names_filename}{the filename to save
the gene IDs and gene names to.
For the experiment, use \code{gene_names.csv}}
}
\value{
names of the \code{rds} files created
}
\description{
For each gene,
create an \code{rds} file containing a tibble of variations per SNP
}
