% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_variations_rds_file.R
\name{create_snp_variations_rds_file}
\alias{create_snp_variations_rds_file}
\title{For each gene's SNPs,
create an \code{rds} file containing a tibble of variations per SNP}
\usage{
create_snp_variations_rds_file(snps_filename, n_snps, verbose = FALSE)
}
\arguments{
\item{snps_filename}{one filenames of a file
containing the SNP IDs,
named \code{[gene_name]_snps.csv}.
These files can be read by \link{read_snps_file}}

\item{n_snps}{the number of SNPs.
Use \link{Inf} to use all SNS IDs}

\item{verbose}{set to TRUE for more output}
}
\value{
names of the \code{rds} files created
}
\description{
For each gene's SNPs,
create an \code{rds} file containing a tibble of variations per SNP
}
