% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_variations_rds_files.R
\name{create_snp_variations_rds_files}
\alias{create_snp_variations_rds_files}
\title{For each gene's SNPs,
create an \code{rds} file containing a tibble of variations per SNP}
\usage{
create_snp_variations_rds_files(snps_filenames, n_snps)
}
\arguments{
\item{snps_filenames}{one or more filenames of files
containing the SNP IDs,
named \code{[gene_name]_snps.csv}.
These files can be read by \link{read_snps_file}}

\item{n_snps}{the number of SNPs.
Use \link{Inf} to use all SNS IDs}
}
\value{
names of the \code{rds} files created
}
\description{
For each gene's SNPs,
create an \code{rds} file containing a tibble of variations per SNP
}
