% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_topo_file.R
\name{create_topo_file}
\alias{create_topo_file}
\title{Create the topology files from the FASTA files}
\usage{
create_topo_file(fasta_filename)
}
\description{
Create the topology files from the FASTA files
}
