% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_topo_file_from_scratch.R
\name{create_topo_file_from_scratch}
\alias{create_topo_file_from_scratch}
\title{Create a \code{.topo} file from scratch}
\usage{
create_topo_file_from_scratch(fasta_filename, verbose = FALSE)
}
\arguments{
\item{fasta_filename}{name of a FASTA file with a protein sequence}

\item{verbose}{set to TRUE for more output}
}
\description{
Create a \code{.topo} file from scratch
}
