% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_topo_files.R
\name{create_topo_files}
\alias{create_topo_files}
\title{Create the topology files from the FASTA files}
\usage{
create_topo_files(fasta_filenames)
}
\description{
Create the topology files from the FASTA files
}
