% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(gene_ids_filename, gene_names_filename, n_gene_ids, verbose)
}
\arguments{
\item{gene_ids_filename}{the filename to save
the gene IDs to.
For the experiment, use \code{gene_ids.csv}}

\item{gene_names_filename}{the filename to save
the gene IDs and gene names to.
For the experiment, use \code{gene_names.csv}}

\item{n_gene_ids}{the number of gene IDs.
Use \link{Inf} to use all gene IDs}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
