% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene_names_file.R
\name{read_gene_names_file}
\alias{read_gene_names_file}
\title{Read teh \code{gene_names.csv} file}
\usage{
read_gene_names_file(gene_names_filename)
}
\arguments{
\item{gene_names_filename}{the filename to save
the gene IDs and gene names to,
as can be read by \link{read_gene_names_file}.
For the experiment, use \code{gene_names.csv}}
}
\description{
Read teh \code{gene_names.csv} file
}
