% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_is_in_tmh_file.R
\name{read_is_in_tmh_file}
\alias{read_is_in_tmh_file}
\title{Read a \code{[gene_name]_is_in_tmh.csv} file}
\usage{
read_is_in_tmh_file(is_in_tmh_filename)
}
\arguments{
\item{is_in_tmh_filename}{the filename to save
the \code{is_in_tmh} table to (as can be read by
\link{read_is_in_tmh_file}).
For the experiment, use \code{[gene_name]_is_in_tmh.csv}}
}
\value{
a \link[tibble]{tibble} with columns named
\code{variation}, \code{is_in_tmh} and \code{p_in_tmh}
}
\description{
Read a \code{[gene_name]_is_in_tmh.csv} file
}
