% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_is_in_tmh_files.R
\name{read_is_in_tmh_files}
\alias{read_is_in_tmh_files}
\title{Read \code{[gene_id]_is_in_tmh.csv} files}
\usage{
read_is_in_tmh_files(is_in_tmh_filenames)
}
\arguments{
\item{is_in_tmh_filenames}{the filename to save
the \code{is_in_tmh} tables to (each of which can be read by
\link{read_is_in_tmh_file}).
For the experiment, use \code{[gene_name]_is_in_tmh.csv}}
}
\description{
Read \code{[gene_id]_is_in_tmh.csv} files
}
