% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results_file.R
\name{read_results_file}
\alias{read_results_file}
\title{Read the results file and removes duplicates}
\usage{
read_results_file(results_filename)
}
\arguments{
\item{results_filename}{name of the results file.
For the experiment, use \code{results.csv}}
}
\description{
Read the results file and removes duplicates
}
