% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snps_file.R
\name{read_snps_file}
\alias{read_snps_file}
\title{Read a \code{[gene_name]_snps.csv} file}
\usage{
read_snps_file(snps_filename)
}
\arguments{
\item{snps_filename}{one filenames of a file
containing the SNP IDs,
named \code{[gene_name]_snps.csv}.
These files can be read by \link{read_snps_file}}
}
\value{
a \link[tibble]{tibble} with one column named \code{snp_id}
}
\description{
Read a \code{[gene_name]_snps.csv} file
}
