% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snps_files.R
\name{read_snps_files}
\alias{read_snps_files}
\title{Read the SNPs filenames to create one big table from it.
To maintain information about the gene name,
the gene names table is used to convert a known gene ID to
a gene name. Note that things will go fine without the gene
names table: the gene name can be deduced later}
\usage{
read_snps_files(snps_filenames)
}
\arguments{
\item{snps_filenames}{one or more filenames of files
containing the SNP IDs,
named \code{[gene_name]_snps.csv}.
These files can be read by \link{read_snps_file}}
}
\description{
Read the SNPs filenames to create one big table from it.
To maintain information about the gene name,
the gene names table is used to convert a known gene ID to
a gene name. Note that things will go fine without the gene
names table: the gene name can be deduced later
}
