% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variations_csv_file.R
\name{read_variations_csv_file}
\alias{read_variations_csv_file}
\title{Read a \code{[gene_name]_variations.csv} file}
\usage{
read_variations_csv_file(variations_csv_filename)
}
\arguments{
\item{variations_csv_filename}{name of a \code{[gene_name]_variations.csv}
file}
}
\value{
a \link[tibble]{tibble} with columns \code{snp_id}
and \code{variation}
}
\description{
Read a \code{[gene_name]_variations.csv} file
}
