% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variations_csv_files.R
\name{read_variations_csv_files}
\alias{read_variations_csv_files}
\title{Read}
\usage{
read_variations_csv_files(variations_csv_filenames)
}
\arguments{
\item{variations_csv_filenames}{names of one or more
\code{[gene_name]_variations.csv}
files}
}
\description{
Read
}
