% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variations_rds_file.R
\name{read_variations_rds_file}
\alias{read_variations_rds_file}
\title{Read a \code{[gene_id]_variations.rds} file}
\usage{
read_variations_rds_file(variations_rds_filename)
}
\arguments{
\item{variations_rds_filename}{name of a \code{[gene_name]_variations.rds}
file}
}
\value{
a list of tibbles
}
\description{
Read a \code{[gene_id]_variations.rds} file
}
