% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_variations_rds_files.R
\name{read_variations_rds_files}
\alias{read_variations_rds_files}
\title{Read one or more \code{[gene_id]_variations.rds} file}
\usage{
read_variations_rds_files(variations_rds_filenames)
}
\arguments{
\item{variations_rds_filenames}{names of one or more
\code{[gene_name]_variations.rds}
files}
}
\value{
a list of lists of tibbles
}
\description{
Read one or more \code{[gene_id]_variations.rds} file
}
