% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_ids.R
\name{create_gene_ids}
\alias{create_gene_ids}
\title{Create all gene IDs of human membrane proteins}
\usage{
create_gene_ids(n_gene_ids, filename)
}
\arguments{
\item{n_gene_ids}{the number of gene IDs.
Use \link{Inf} to use all gene IDs}
}
\description{
Create all gene IDs of human membrane proteins
}
