% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_is_in_tmh_files.R
\name{create_is_in_tmh_files}
\alias{create_is_in_tmh_files}
\title{Create the \code{[gene_name]_is_in_tmh.csv} files.}
\usage{
create_is_in_tmh_files(variations_csv_filenames, topo_filenames)
}
\arguments{
\item{variations_csv_filenames}{names of one or more
\code{[gene_name]_variations.csv}
files}

\item{topo_filenames}{name of one or more \code{.topo} files
with protein topologies}
}
\description{
Create the \code{[gene_name]_is_in_tmh.csv} files.
}
