% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_results_file.R
\name{create_results_file}
\alias{create_results_file}
\title{Create the results file}
\usage{
create_results_file(
  gene_names_filename = "gene_names.csv",
  snp_ids_filenames,
  variations_csv_filenames,
  is_in_tmh_filenames,
  results_filename = "results.csv"
)
}
\arguments{
\item{gene_names_filename}{the filename to save
the gene IDs and gene names to,
as can be read by \link{read_gene_names_file}.
For the experiment, use \code{gene_names.csv}}

\item{variations_csv_filenames}{names of one or more
\code{[gene_name]_variations.csv}
files}

\item{is_in_tmh_filenames}{the filename to save
the \code{is_in_tmh} tables to (each of which can be read by
\link{read_is_in_tmh_file}).
For the experiment, use \code{[gene_name]_is_in_tmh.csv}}

\item{results_filename}{name of the results file.
For the experiment, use \code{results.csv}}
}
\description{
Create the results file
}
